Fixing Bifurcation diagram
There was an issue of using a folder of different folder structure than desired leading to not being able to create the bifurcation diagram. We fixed this now.
Furthermore, I changed the labelling/coloring of the bifurcation diagram and made it easier to manually correct the misclassified data for the bifurcation diagram.